# PUBDEV-5760: Avoid unnecessary memory allocation in GBM/DRF mojo scoring