--- a/h2o-genmodel/src/main/java/hex/genmodel/algos/tree/ScoreTree.java
+++ b/h2o-genmodel/src/main/java/hex/genmodel/algos/tree/ScoreTree.java
@@ -1,0 +1,9 @@
+package hex.genmodel.algos.tree;
+
+import java.io.Serializable;
+
+public interface ScoreTree extends Serializable {
+
+  double scoreTree(byte[] tree, double[] row, int nclasses, boolean computeLeafAssignment, String[][] domains);
+
+}