--- a/h2o-genmodel/src/main/java/hex/genmodel/algos/tree/ScoreTree1.java
+++ b/h2o-genmodel/src/main/java/hex/genmodel/algos/tree/ScoreTree1.java
@@ -1,0 +1,10 @@
+package hex.genmodel.algos.tree;
+
+public final class ScoreTree1 implements ScoreTree {
+
+  @Override
+  public final double scoreTree(byte[] tree, double[] row, int nclasses, boolean computeLeafAssignment, String[][] domains) {
+    return SharedTreeMojoModel.scoreTree1(tree, row, nclasses, computeLeafAssignment);
+  }
+
+}