Field _mojo_version was declared as Number, comparison `_mojo_version.equals(1.0)`
is harmful because it triggers autoboxing of a constant to an object.
This happens for every scored row.

Microbenchmark results:

- Current (master) version:

     Benchmark                             (rows)  Mode  Cnt    Score   Error  Units
     GbmMojoScoringBench.measureGbmScore0    1000  avgt   10    5.805 ± 0.044  ms/op
     GbmMojoScoringBench.measureGbmScore0  100000  avgt   10  571.884 ± 3.827  ms/op

- This PR:

     Benchmark                             (rows)  Mode  Cnt    Score   Error  Units
     GbmMojoScoringBench.measureGbmScore0    1000  avgt   10    5.643 ± 0.039  ms/op
     GbmMojoScoringBench.measureGbmScore0  100000  avgt   10  555.807 ± 5.626  ms/op

- Optimal version (simplified implementation that assumes single MOJO version and doesn't need to make a decision based on version):

     Benchmark                             (rows)  Mode  Cnt    Score   Error  Units
     GbmMojoScoringBench.measureGbmScore0    1000  avgt   10    5.658 ± 0.059  ms/op
     GbmMojoScoringBench.measureGbmScore0  100000  avgt   10  558.564 ± 3.152  ms/op

This PR improves the performance of scoring, it is now similar to the "optimal" version. There is no unnecessary object allocation.

To reproduce run:

    gw -PdoUBench=true -PubenchIncludeOnly=GbmMojoScoringBench :h2o-algos:ubench -x test